Backend: Refactor to prepare for adding more backends
