+++ llvmexpr/runtime/llvm/Compiler.cpp
@@ -35,9 +35,9 @@
 #include "llvm/Support/FileSystem.h"
 #include "llvm/Support/raw_ostream.h"
 
-#include "../ir/ExprIRGenerator.hpp"
-#include "../ir/SingleExprIRGenerator.hpp"
-#include "../utils/Diagnostics.hpp"
+#include "../../codegen/llvm/Diagnostics.hpp"
+#include "../../codegen/llvm/LLVMGenerator.hpp"
+#include "../../codegen/llvm/SingleExprIRGenerator.hpp"
 
 Compiler::Compiler(
     std::vector<Token> tokens_in, const VSVideoInfo* out_vi,
@@ -103,7 +103,7 @@ CompiledFunction Compiler::compile_with_approx_math(int actual_approx_math) {
     // Create IR generator and generate code
     std::unique_ptr<IRGeneratorBase> ir_gen;
     if (expr_mode == ExprMode::EXPR) {
-        ir_gen = std::make_unique<ExprIRGenerator>(
+        ir_gen = std::make_unique<LLVMGenerator>(
             tokens, vo, vi, width, height, mirror_boundary, prop_map,
             analysis_results, *context, *module, builder, math_manager,
             func_name, actual_approx_math);
